// src/components/tabs-progress-report/models/state-progress-model.ts
export interface StateProgressModel {
  descripAmount: string;
  quantityTotalCost: number;
  percentTotalCost: number;
  quantityProgressToday: number;
  percentProgressToday: number;
  quantityLastProgress: number;
  percentLastProgress: number;
  quantityProgressPeriod: number;
  percentProgressPeriod: number;
}